requirements.txt
PyQt6==6.7.0
colorlog
pyqtgraph
pyzmq
open3d
trimesh
numpy
shapely
mapbox_earcut
lxml
networkx
pycollada